[Refactor] : 검색결과페이지 리팩토링 - layout.tsx를 만들어 여기서 Suspense를 감싸줌 - SearchResult.tsx를 pages.tsx로 옮겨줌 - API 요청, 필터링 관련 로직을 hook으로 분리
