+++ src/app/(NavBarCommonLayout)/search/result/layout.tsx
@@ -1,0 +1,6 @@
+import TigLoadingPage from '@components/all/TigLoadingPage';
+import { PropsWithChildren, Suspense } from 'react';
+
+export default function Layout({ children }: PropsWithChildren) {
+  return <Suspense fallback={<TigLoadingPage />}>{children}</Suspense>;
+}
